+++ test/erc20_gateway/erc20_cogateway/confirm_deposit.js
@@ -1,3 +1,5 @@
+'use-strict';
+
 // Copyright 2020 OpenST Ltd.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
@@ -11,8 +13,6 @@
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
-
-'use-strict';
 
 const BN = require('bn.js');
